@@ -1,15 +1,15 @@
-<?php namespace Sharenjoy\Cmsharenjoy\Repo\Order;
+<?php namespace Sharenjoy\Cmsharenjoy\Repo\Qna;
 
 use Sharenjoy\Cmsharenjoy\Core\EloquentBaseRepository;
 use Sharenjoy\Cmsharenjoy\Service\Validation\ValidableInterface;
 
-class OrderRepository extends EloquentBaseRepository implements OrderInterface {
+class QnaRepository extends EloquentBaseRepository implements QnaInterface {
     
-    public function __construct(Order $model, ValidableInterface $validator)
+    public function __construct(Qna $model, ValidableInterface $validator)
     {
         $this->validator = $validator;
         $this->model     = $model;
-
+        
         parent::__construct();
     }
 
