@@ -12,27 +12,19 @@ class UserTable extends Seeder {
 
         $types = [
             [
-                'email'         => Config::get('cmsharenjoy::setup.email'),
-                'password'      => Hash::make(Config::get('cmsharenjoy::setup.password')),
+                'email'         => Config::get('cmsharenjoy::app.administrator.email'),
+                'password'      => Hash::make(Config::get('cmsharenjoy::app.administrator.password')),
                 'activated'     => '1',
+                'activated_at'  => date('Y-m-d H:i:s'),
+                'name'          => Config::get('cmsharenjoy::app.administrator.name'),
+                'phone'         => Config::get('cmsharenjoy::app.administrator.phone'),
+                'sort'          => time(),
                 'created_at'    => date('Y-m-d H:i:s'),
                 'updated_at'    => date('Y-m-d H:i:s')
             ]
         ];
         DB::table('users')->insert($types);
         $this->command->info('User Table Seeded');
-
-        $types = [
-            [
-                'user_id'       => '1',
-                'first_name'    => Config::get('cmsharenjoy::setup.first_name'),
-                'last_name'     => Config::get('cmsharenjoy::setup.last_name'),
-                'phone'         => Config::get('cmsharenjoy::setup.phone')
-            ]
-        ];
-        DB::table('accounts')->insert($types);
-        $this->command->info('Account Table Seeded');
-
     }
 
 }
