@@ -1,11 +1,9 @@
 <?php namespace Sharenjoy\Cmsharenjoy\Controllers;
 
 use Sharenjoy\Cmsharenjoy\Setting\SettingInterface;
-use Input, Redirect, Message, Debugbar;
+use Input, Redirect, Message, Form, Response, Request;
 
 class SettingController extends BaseController {
-
-    protected $appName = 'setting';
 
     public function __construct(SettingInterface $setting)
     {
@@ -15,41 +13,53 @@ class SettingController extends BaseController {
 
     public function getIndex()
     {
-        $model = $this->repository->getModel();
-        $items = $this->item($model);
+        $items = $this->item();
         
-        return $this->layout->with('items', $items);
+        $buttons = Form::button(trans('cmsharenjoy::buttons.save'), array('class'=>'btn btn-success btn-save')).'&nbsp;'.
+                   Form::button(trans('cmsharenjoy::buttons.reset'), array('class'=>'btn btn-reset'));
+
+        return $this->layout->with('items', $items)
+                            ->with('buttons', $buttons);
     }
 
-    public function postIndex()
+    public function postStore()
     {
-        $settings = Input::all();
-        $settings = array_except($settings, array('_token', 'type'));
+        if( ! Request::ajax()) Response::json('error', 400);
 
-        $model = $this->repository->getModel();
+        $data['id']    = Input::get('id');
+        $data['item']  = Input::get('item');
+        $data['value'] = Input::get('value');
 
-        foreach($settings as $key => $value)
+        if (Input::get('type') == 'checkbox' && $data['value'] != '')
         {
-            $model = $model->where('key', $key)->first();
-
-            if ($model)
+            $array = explode('&', $data['value']);
+            foreach($array as $value)
             {
-                $model->value = $value;
-                $model->save();
+                $val = explode('=', $value);
+                $ary[] = $val[1];
             }
+            $data['value'] = implode(',', $ary);
         }
 
-        $items = $this->item($model);
+        try
+        {
+            $this->repository->edit($data['item'], array('value' => $data['value']), 'key');
+        }
+        catch (\Sharenjoy\Cmsharenjoy\Exception\EntityNotFoundException $e)
+        {
+            Message::output('flash', 'errors', trans('cmsharenjoy::exception.not_found', array('id' => $data['id'])));
+            return Redirect::to($this->objectUrl);
+        }
 
-        Message::add('success', trans('cmsharenjoy::admin.success_updated'))->flash();
-        
-        return Redirect::to($this->objectUrl)->with('items', $items);
+        return Response::json(Message::output('json', 'success', trans('cmsharenjoy::admin.success_ordered'), $data), 200);
     }
 
-    protected function item($model)
+    protected function item()
     {
-        $items['general'] = $model->where('module', 'general')->get();
-        $items['file']    = $model->where('module', 'file')->get();
+        $model = $this->repository->getModel();
+
+        $items['general']['item'] = $model->where('module', 'general')->get();
+        $items['file']['item']    = $model->where('module', 'file')->get();
 
         return $items;
     }
