@@ -1,7 +1,27 @@
 <?php
 
 return array(
-    'pleaseSelect'                        => '請選擇',
-    'enable'                              => '開啟',
-    'disable'                             => '關閉'
+
+    // status
+    'pleaseSelect'     => '請選擇',
+    'enable'           => '開啟',
+    'disable'          => '關閉',
+    
+    // process
+    'process1'         => '訂單確認中',
+    'process2'         => '訂單準備出貨中',
+    'process3'         => '訂單已出貨寄送中',
+    'process4'         => '訂單寄送完成已送達',
+    'process5'         => '訂單退款中',
+    'process6'         => '訂單取消',
+
+    // easy_contact_time
+    'easyContactTime1' => '任何時間皆可',
+    'easyContactTime2' => '上午09:00-11:00',
+    'easyContactTime3' => '中午11:00-14:00',
+    'easyContactTime4' => '下午14:00-18:00',
+    'easyContactTime5' => '晚上18:00-22:00',
+    
+    
+
 );
