@@ -19,11 +19,10 @@ class CreateSettingsTable extends Migration {
 
                 $table->increments('id')->index();
                 $table->string('key', 255)->unique();
-                $table->string('label', 255);
-                $table->text('description')->nullable();
+                $table->string('module', 255);
                 $table->string('type', 255);
                 $table->string('value', 255);
-                $table->string('module', 255);
+                $table->text('option')->nullable();
                 $table->integer('sort')->unsigned()->nullable()->default(0);
                 $table->timestamps();
             });
